+++ nix/pkgs.nix
@@ -34,6 +34,20 @@
           ];
         };
 
+        grace = super.ocamlPackages.buildDunePackage {
+          pname = "grace";
+          version = "0.2.0";
+          src = self.sources.grace-src;
+          buildInputs = with super.ocamlPackages; [
+            dedent
+            core
+            ppx_jane
+            iter
+            uutf
+            fmt
+          ];
+        };
+
         # js_of_ocaml-compiler version 6.0.1 is misbehaving
         js_of_ocaml-compiler = super.ocamlPackages.js_of_ocaml-compiler.override { version = "5.9.1"; };
         js_of_ocaml = super.ocamlPackages.js_of_ocaml.override { inherit js_of_ocaml-compiler; };