+++ src/ir_def/check_ir.ml
@@ -117,7 +117,7 @@ let lub env t1 t2 =
 
 exception CheckFailed of string
 
-let type_error at : string -> Diag.message = Diag.error_message at "M0000" "IR type"
+let type_error at : string -> Diag.message = Diag.error_message at "M0000" "IR type" ~notes:[] ~spans:[]
 
 let error env at fmt =
     Printf.ksprintf (fun s -> raise (CheckFailed (Diag.string_of_message (type_error at s)))) fmt