+++ src/mo_config/args.ml
@@ -29,9 +29,10 @@ let package_args = [
 let error_args = [
   "--error-detail", Arg.Set_int Flags.error_detail, "<n>  set error message detail for syntax errors, n in [0..3] (default 2)";
   "--error-recovery", Arg.Set Flags.error_recovery, " report multiple syntax errors";
-  "--error-format",     Arg.Symbol (["plain"; "json"], fun s ->
+  "--error-format",     Arg.Symbol (["plain"; "human"; "json"], fun s ->
       Flags.error_format := (match s with
         | "json" -> Flags.Json
+        | "human" -> Flags.Human
         | _ -> Flags.Plain)),
     " set error output format"
   (* TODO move --hide-warnings here? *)