+++ src/mo_frontend/typing.ml
@@ -249,11 +249,11 @@ let display_typs fmt typs =
   else
     Format.fprintf fmt ""
 
-let type_error at code text : Diag.message =
-  Diag.error_message at code "type" text
+let type_error at code text notes spans : Diag.message =
+  Diag.error_message at code "type" text ~notes ~spans
 
-let type_warning at code text : Diag.message =
-  Diag.warning_message at code "type" text
+let type_warning at code text notes spans : Diag.message =
+  Diag.warning_message at code "type" text ~notes ~spans
 
 let type_info at text : Diag.message =
   Diag.info_message at "type" text
@@ -271,20 +271,20 @@ module Format = struct
   let fprintf env = with_con_map env Format.fprintf
 end
 
-let error env at code fmt =
+let error ?(notes = []) ?(spans = []) env at code fmt =
   Format.kasprintf env (fun s ->
-      Diag.add_msg env.msgs (type_error at code s);
+      Diag.add_msg env.msgs (type_error at code s notes spans);
       raise Recover)
     fmt
 
-let local_error env at code fmt =
+let local_error ?(notes = []) ?(spans = []) env at code fmt =
   Format.kasprintf env (fun s ->
-      Diag.add_msg env.msgs (type_error at code s))
+      Diag.add_msg env.msgs (type_error at code s notes spans))
     fmt
 
-let warn env at code fmt =
+let warn ?(notes = []) ?(spans = []) env at code fmt =
   Format.kasprintf env (fun s ->
-      if not env.errors_only then Diag.add_msg env.msgs (type_warning at code s))
+      if not env.errors_only then Diag.add_msg env.msgs (type_warning at code s notes spans))
     fmt
 
 let info env at fmt =
@@ -303,7 +303,7 @@ let check_deprecation env at desc id depr =
       (match compare !Flags.experimental_stable_memory 0 with
        | -1 -> error
        | 0 -> warn
-       | _ -> fun _ _ _ _ -> ())
+       | _ -> fun ?(notes = []) ?(spans = []) _ _ _ _ -> ())
        env at code
        "this code is (or uses) the deprecated library `ExperimentalStableMemory`.\nPlease use the `Region` library instead: https://internetcomputer.org/docs/current/motoko/main/stable-memory/stable-regions/#the-region-library or compile with flag `--experimental-stable-memory 1` to suppress this message."
     end
@@ -318,7 +318,7 @@ let flag_of_compile_mode mode =
   | Flags.WasmMode -> " and flag -no-system-api"
   | Flags.RefMode -> " and flag -ref-system-api"
 
-let diag_in type_diag modes env at code fmt =
+let diag_in type_diag modes env at code notes spans fmt =
   let mode = !Flags.compile_mode in
   if !Flags.compiled && List.mem mode modes then
     begin
@@ -329,13 +329,13 @@ let diag_in type_diag modes env at code fmt =
             s
             (flag_of_compile_mode mode)
           in
-          Diag.add_msg env.msgs (type_diag at code s)) fmt;
+          Diag.add_msg env.msgs (type_diag at code s notes spans)) fmt;
       true
     end
   else false
 
-let error_in modes env at code fmt =
-  if diag_in type_error modes env at code fmt then
+let error_in modes env at code ?(notes=[]) ?(spans=[]) fmt =
+  if diag_in type_error modes env at code notes spans fmt then
     raise Recover
 
 let plural cs = if T.ConSet.cardinal cs = 1 then "" else "s"
@@ -349,8 +349,8 @@ let warn_lossy_bind_type env at bind t1 t2 =
       display_typ_expand t2
 
 (* Currently unused *)
-let _warn_in modes env at code fmt =
-  ignore (diag_in type_warning modes env at code fmt)
+let _warn_in modes env at code notes spans fmt =
+  ignore (diag_in type_warning modes env at code notes spans fmt)
 
 (* Unused identifier detection *)
 
@@ -361,15 +361,25 @@ let emit_unused_warnings env =
   let emit (id, region, kind) =
     if is_in_shared_pat region.left then
       match kind with
-      | Scope.Declaration -> warn env region "M0240" "unused identifier %s in shared pattern (delete or rename to wildcard `_` or `_%s`)" id id
-      | Scope.FieldReference -> warn env region "M0241" "unused field %s in shared pattern (delete or rewrite as `%s = _`)" id id
-      | Scope.MutableNotAssigned -> warn env region "M0244" "variable %s is never reassigned, consider using `let`" id
+      | Scope.Declaration ->
+         warn env region "M0240"
+           "unused identifier %s in shared pattern (delete or rename to wildcard `_` or `_%s`)" id id
+      | Scope.FieldReference ->
+         warn env region "M0241"
+           "unused field %s in shared pattern (delete or rewrite as `%s = _`)" id id
+      | Scope.MutableNotAssigned ->
+         warn env region "M0244" "variable %s is never reassigned, consider using `let`" id
       | Scope.MixinIncluded -> ()
     else
       match kind with
-      | Scope.Declaration -> warn env region "M0194" "unused identifier %s (delete or rename to wildcard `_` or `_%s`)" id id
-      | Scope.FieldReference -> warn env region "M0198" "unused field %s in object pattern (delete or rewrite as `%s = _`)" id id
-      | Scope.MutableNotAssigned -> warn env region "M0244" "variable %s is never reassigned, consider using `let`" id
+      | Scope.Declaration ->
+        warn env region  "M0194"
+          "unused identifier %s (delete or rename to wildcard `_` or `_%s`)" id id
+      | Scope.FieldReference ->
+        warn env region "M0198"
+          "unused field %s in object pattern (delete or rewrite as `%s = _`)" id id
+      | Scope.MutableNotAssigned ->
+         warn env region "M0244" "variable %s is never reassigned, consider using `let`" id
       | Scope.MixinIncluded -> ()
   in
   UWSet.iter emit !(env.unused_warnings)
@@ -474,7 +484,7 @@ let coverage' warnOrError category env f x t at =
   let uncovered, unreached = f x t in
   List.iter (fun at -> warn env at "M0146" "this pattern is never matched") unreached;
   if uncovered <> [] then
-    warnOrError env at "M0145"
+    warnOrError ?notes:None ?spans:None env at "M0145"
       ("this %s of type%a\ndoes not cover value\n  %s" : (_, _, _, _) format4 )
       category
       display_typ_expand t
@@ -642,7 +652,7 @@ let error_shared env t at code fmt =
         display_typ_expand t
         display_typ_expand t1
     in
-    Format.kasprintf env (fun s1 -> Diag.add_msg env.msgs (type_error at code (s1^s)); raise Recover) fmt
+    Format.kasprintf env (fun s1 -> Diag.add_msg env.msgs (type_error at code (s1^s) [] []); raise Recover) fmt
 
 let as_domT t =
   match t.Source.it with
@@ -1565,22 +1575,20 @@ let resolve_hole env at hole_sort typ =
   | [] ->
     let (lib_terms, _) = candidates true env.libs is_lib_module in
     (match if Option.is_some !Flags.implicit_package then disambiguate_holes lib_terms else None with
-    | Some term -> Ok term
-    | None -> Error (HoleSuggestions (lib_terms, explicit_terms, renaming_hints)))
-  | terms -> begin
+      | Some term -> Ok term
+      | None -> Error (HoleSuggestions (lib_terms, explicit_terms, renaming_hints)))
+  | terms ->
      match disambiguate_holes terms with
      | Some term -> Ok term
      | None -> Error (HoleAmbiguous (fun env ->
        let terms = List.map desc_of_candidate terms in
-       error env at "M0231" "ambiguous implicit argument %s of type%a.\nThe ambiguous implicit candidates are: %s%s."
+       let notes = Printf.sprintf "The ambiguous implicit candidates are: %s." (String.concat ", " terms) ::
+         if explicit_terms = [] then [] else
+            [ "The other explicit candidates are: " ^ (String.concat ", " (List.map desc_of_candidate explicit_terms)) ]
+       in
+       error env at "M0231" ~notes "ambiguous implicit argument %s of type %a."
          (match hole_sort with Named n -> "named " ^ quote n | Anon i -> "at argument position " ^ Int.to_string i)
-         display_typ typ
-         (String.concat ", " terms)
-         (if explicit_terms = [] then ""
-          else
-            ".\nThe other explicit candidates are: "^
-              (String.concat ", " (List.map desc_of_candidate explicit_terms)))))
-     end
+         display_typ typ))
 
 type ctx_dot_candidate =
   { module_ref : T.lab option; (* optional module reference : name (from `vals`) or path (from `libs`) *)
@@ -2319,7 +2327,7 @@ and try_infer_dot_exp env at exp id (desc, pred) =
         type_error exp.at "M0070"
           (Format.asprintf env
              "expected object type, but expression produces type%a"
-             display_typ_expand t0))
+             display_typ_expand t0) [] [])
   in
   match fields with
   | Error e -> Error e
@@ -2327,31 +2335,31 @@ and try_infer_dot_exp env at exp id (desc, pred) =
     let suggest () =
       Suggest.suggest_id "field" id.it (List.map (fun f -> f.T.lab) fs)
     in
-    match T.lookup_val_field id.it fs with
-    | T.Pre ->
+    match T.lookup_val_field_opt id.it fs with
+    | Some(T.Pre) ->
       error env at "M0071"
         "cannot infer type of forward field reference %s"
         id.it
-    | t when pred (T.as_immut t) ->
+    | Some(t) when pred (T.as_immut t) ->
       if not env.pre then
         check_deprecation env at "field" id.it (T.lookup_val_deprecation id.it fs);
       Ok(t)
-    | t (* when not (pred t) *) ->
+    | Some(t) (* when not (pred t) *) ->
       Error(t1, fun () ->
         type_error id.at "M0234"
           (Format.asprintf env "field %s does exist in %a\nbut is not %s.\n%s"
              id.it
              display_obj t0
              desc
-             (suggest ())))
-    | exception Invalid_argument _ ->
+             (suggest ())) [] [])
+    | None ->
       Error(t1, fun () ->
         type_error id.at "M0072"
           (Format.asprintf env "field %s does not exist in %a%s"
              id.it
              display_obj t0
              (Suggest.suggest_id "field" id.it
-                (List.map (fun f -> f.T.lab) fs))))
+                (List.map (fun f -> f.T.lab) fs))) [] [])
     end
 
 and infer_exp_field env rf =
@@ -2454,24 +2462,23 @@ and check_exp' env0 t exp : T.typ =
       t
     | Error (HoleSuggestions (lib_terms, explicit_terms, renaming_hints)) ->
       if not env.pre then begin
+        let explicit_sug =
+          if explicit_terms = [] then []
+          else [Printf.sprintf "Did you mean to explicitly use %s?" (String.concat " or " (List.map desc_of_candidate explicit_terms))]
+        in
         let import_sug =
           if lib_terms = [] then
             let desc = match s with Named id -> " named " ^ quote id | _ -> "" in
             Stdlib.Format.sprintf
-             "\nHint: If you're trying to omit an implicit argument%s you need to have a matching declaration%s in scope."
+             "If you're trying to omit an implicit argument%s you need to have a matching declaration%s in scope."
              desc desc
-          else Stdlib.Format.sprintf "\nHint: Did you mean to import %s?" (String.concat " or " (List.filter_map import_suggestion_of_candidate lib_terms))
-        in
-        let explicit_sug =
-          if explicit_terms = [] then ""
-          else Stdlib.Format.sprintf "\nHint: Did you mean to explicitly use %s?" (String.concat " or " (List.map desc_of_candidate explicit_terms))
+          else Stdlib.Format.sprintf "Did you mean to import %s?" (String.concat " or " (List.filter_map import_suggestion_of_candidate lib_terms))
         in
         renaming_hints env;
-        local_error env exp.at "M0230" "Cannot determine implicit argument %s of type%a%s%s"
+        let notes = import_sug::explicit_sug in
+        local_error ~notes env exp.at "M0230" "Cannot determine implicit argument %s of type%a"
           (desc s)
           display_typ t
-          import_sug
-          explicit_sug
       end;
       t
   end